+++ args_test.go
@@ -6,6 +6,7 @@ package argument_test
 
 import (
 	"bytes"
+	"context"
 	"flag"
 	"os"
 	"time"
@@ -17,28 +18,30 @@ import (
 )
 
 var _ = Describe("ParseArgs", func() {
+	var ctx context.Context
 	BeforeEach(func() {
+		ctx = context.Background()
 		flag.CommandLine.SetOutput(&bytes.Buffer{})
 		flag.CommandLine = flag.NewFlagSet(os.Args[0], flag.ContinueOnError)
 	})
 	It("parse empty struct", func() {
 		var args struct {
 		}
-		err := argument.ParseArgs(&args, []string{})
+		err := argument.ParseArgs(ctx, &args, []string{})
 		Expect(err).To(BeNil())
 	})
 	It("ignore private fields", func() {
 		var args struct {
 			private string
 		}
-		err := argument.ParseArgs(&args, []string{})
+		err := argument.ParseArgs(ctx, &args, []string{})
 		Expect(err).To(BeNil())
 	})
 	It("parse string from args parameter", func() {
 		var args struct {
 			Username string `arg:"user"`
 		}
-		err := argument.ParseArgs(&args, []string{"-user=Ben"})
+		err := argument.ParseArgs(ctx, &args, []string{"-user=Ben"})
 		Expect(err).NotTo(HaveOccurred())
 		Expect(args.Username).To(Equal("Ben"))
 	})
@@ -46,7 +49,7 @@ var _ = Describe("ParseArgs", func() {
 		var args struct {
 			Username string `arg:"user" default:"Ben"`
 		}
-		err := argument.ParseArgs(&args, []string{})
+		err := argument.ParseArgs(ctx, &args, []string{})
 		Expect(err).NotTo(HaveOccurred())
 		Expect(args.Username).To(Equal("Ben"))
 	})
@@ -54,7 +57,7 @@ var _ = Describe("ParseArgs", func() {
 		var args struct {
 			Age int `arg:"age"`
 		}
-		err := argument.ParseArgs(&args, []string{"-age=29"})
+		err := argument.ParseArgs(ctx, &args, []string{"-age=29"})
 		Expect(err).NotTo(HaveOccurred())
 		Expect(args.Age).To(Equal(29))
 	})
@@ -62,21 +65,21 @@ var _ = Describe("ParseArgs", func() {
 		var args struct {
 			Age int `arg:"age"`
 		}
-		err := argument.ParseArgs(&args, []string{"-age=abc"})
+		err := argument.ParseArgs(ctx, &args, []string{"-age=abc"})
 		Expect(err).To(HaveOccurred())
 	})
 	It("return error if parse fails", func() {
 		var args struct {
 			Age int `arg:"age"`
 		}
-		err := argument.ParseArgs(&args, []string{"-age=abc"})
+		err := argument.ParseArgs(ctx, &args, []string{"-age=abc"})
 		Expect(err).To(HaveOccurred())
 	})
 	It("skip fields without tag", func() {
 		var args struct {
 			Age int
 		}
-		err := argument.ParseArgs(&args, []string{})
+		err := argument.ParseArgs(ctx, &args, []string{})
 		Expect(err).NotTo(HaveOccurred())
 		Expect(args.Age).To(Equal(0))
 	})
@@ -84,7 +87,7 @@ var _ = Describe("ParseArgs", func() {
 		var args struct {
 			Age int `arg:"age" default:"29"`
 		}
-		err := argument.ParseArgs(&args, []string{})
+		err := argument.ParseArgs(ctx, &args, []string{})
 		Expect(err).NotTo(HaveOccurred())
 		Expect(args.Age).To(Equal(29))
 	})
@@ -92,7 +95,7 @@ var _ = Describe("ParseArgs", func() {
 		var args struct {
 			Age int64 `arg:"age"`
 		}
-		err := argument.ParseArgs(&args, []string{"-age=29"})
+		err := argument.ParseArgs(ctx, &args, []string{"-age=29"})
 		Expect(err).NotTo(HaveOccurred())
 		Expect(args.Age).To(Equal(int64(29)))
 	})
@@ -100,7 +103,7 @@ var _ = Describe("ParseArgs", func() {
 		var args struct {
 			Age int64 `arg:"age" default:"29"`
 		}
-		err := argument.ParseArgs(&args, []string{})
+		err := argument.ParseArgs(ctx, &args, []string{})
 		Expect(err).NotTo(HaveOccurred())
 		Expect(args.Age).To(Equal(int64(29)))
 	})
@@ -108,7 +111,7 @@ var _ = Describe("ParseArgs", func() {
 		var args struct {
 			Confirm bool `arg:"confirm"`
 		}
-		err := argument.ParseArgs(&args, []string{"-confirm=true"})
+		err := argument.ParseArgs(ctx, &args, []string{"-confirm=true"})
 		Expect(err).NotTo(HaveOccurred())
 		Expect(args.Confirm).To(BeTrue())
 	})
@@ -116,7 +119,7 @@ var _ = Describe("ParseArgs", func() {
 		var args struct {
 			Confirm bool `arg:"confirm"`
 		}
-		err := argument.ParseArgs(&args, []string{"-confirm=false"})
+		err := argument.ParseArgs(ctx, &args, []string{"-confirm=false"})
 		Expect(err).NotTo(HaveOccurred())
 		Expect(args.Confirm).To(BeFalse())
 	})
@@ -124,7 +127,7 @@ var _ = Describe("ParseArgs", func() {
 		var args struct {
 			Confirm bool `arg:"confirm" default:"true"`
 		}
-		err := argument.ParseArgs(&args, []string{})
+		err := argument.ParseArgs(ctx, &args, []string{})
 		Expect(err).NotTo(HaveOccurred())
 		Expect(args.Confirm).To(BeTrue())
 	})
@@ -132,7 +135,7 @@ var _ = Describe("ParseArgs", func() {
 		var args struct {
 			Confirm bool `arg:"confirm" default:"false"`
 		}
-		err := argument.ParseArgs(&args, []string{})
+		err := argument.ParseArgs(ctx, &args, []string{})
 		Expect(err).NotTo(HaveOccurred())
 		Expect(args.Confirm).To(BeFalse())
 	})
@@ -140,14 +143,14 @@ var _ = Describe("ParseArgs", func() {
 		var args struct {
 			Confirm bool `arg:"confirm"`
 		}
-		err := argument.ParseArgs(&args, []string{"-confirm=banana"})
+		err := argument.ParseArgs(ctx, &args, []string{"-confirm=banana"})
 		Expect(err).To(HaveOccurred())
 	})
 	It("parse duration", func() {
 		var args struct {
 			Wait time.Duration `arg:"wait"`
 		}
-		err := argument.ParseArgs(&args, []string{"-wait=1m"})
+		err := argument.ParseArgs(ctx, &args, []string{"-wait=1m"})
 		Expect(err).NotTo(HaveOccurred())
 		Expect(args.Wait).To(Equal(time.Minute))
 	})
@@ -155,15 +158,31 @@ var _ = Describe("ParseArgs", func() {
 		var args struct {
 			Wait time.Duration `arg:"wait" default:"1m"`
 		}
-		err := argument.ParseArgs(&args, []string{})
+		err := argument.ParseArgs(ctx, &args, []string{})
 		Expect(err).NotTo(HaveOccurred())
 		Expect(args.Wait).To(Equal(time.Minute))
+	})
+	It("parse duration days", func() {
+		var args struct {
+			Wait time.Duration `arg:"wait"`
+		}
+		err := argument.ParseArgs(ctx, &args, []string{"-wait=7d"})
+		Expect(err).NotTo(HaveOccurred())
+		Expect(args.Wait).To(Equal(7 * 24 * time.Hour))
+	})
+	It("default duration days", func() {
+		var args struct {
+			Wait time.Duration `arg:"wait" default:"7d"`
+		}
+		err := argument.ParseArgs(ctx, &args, []string{})
+		Expect(err).NotTo(HaveOccurred())
+		Expect(args.Wait).To(Equal(7 * 24 * time.Hour))
 	})
 	It("parse float64", func() {
 		var args struct {
 			Age float64 `arg:"age"`
 		}
-		err := argument.ParseArgs(&args, []string{"-age=29"})
+		err := argument.ParseArgs(ctx, &args, []string{"-age=29"})
 		Expect(err).NotTo(HaveOccurred())
 		Expect(args.Age).To(Equal(float64(29)))
 	})
@@ -171,7 +190,7 @@ var _ = Describe("ParseArgs", func() {
 		var args struct {
 			Age *float64 `arg:"age"`
 		}
-		err := argument.ParseArgs(&args, []string{"-age=29"})
+		err := argument.ParseArgs(ctx, &args, []string{"-age=29"})
 		Expect(err).NotTo(HaveOccurred())
 		Expect(*args.Age).To(Equal(float64(29)))
 	})
@@ -179,7 +198,7 @@ var _ = Describe("ParseArgs", func() {
 		var args struct {
 			Age float64 `arg:"age" default:"29"`
 		}
-		err := argument.ParseArgs(&args, []string{})
+		err := argument.ParseArgs(ctx, &args, []string{})
 		Expect(err).NotTo(HaveOccurred())
 		Expect(args.Age).To(Equal(float64(29)))
 	})
@@ -187,7 +206,7 @@ var _ = Describe("ParseArgs", func() {
 		var args struct {
 			Age *float64 `arg:"age"`
 		}
-		err := argument.ParseArgs(&args, []string{"-age="})
+		err := argument.ParseArgs(ctx, &args, []string{"-age="})
 		Expect(err).NotTo(HaveOccurred())
 		Expect(args.Age).To(BeNil())
 	})
@@ -195,7 +214,7 @@ var _ = Describe("ParseArgs", func() {
 		var args struct {
 			Age *float64 `arg:"age" default:"29"`
 		}
-		err := argument.ParseArgs(&args, []string{})
+		err := argument.ParseArgs(ctx, &args, []string{})
 		Expect(err).NotTo(HaveOccurred())
 		Expect(*args.Age).To(Equal(float64(29)))
 	})
@@ -204,7 +223,7 @@ var _ = Describe("ParseArgs", func() {
 		var args struct {
 			Age uint `arg:"age"`
 		}
-		err := argument.ParseArgs(&args, []string{"-age=29"})
+		err := argument.ParseArgs(ctx, &args, []string{"-age=29"})
 		Expect(err).NotTo(HaveOccurred())
 		Expect(args.Age).To(Equal(uint(29)))
 	})
@@ -212,7 +231,7 @@ var _ = Describe("ParseArgs", func() {
 		var args struct {
 			Age uint `arg:"age" default:"29"`
 		}
-		err := argument.ParseArgs(&args, []string{})
+		err := argument.ParseArgs(ctx, &args, []string{})
 		Expect(err).NotTo(HaveOccurred())
 		Expect(args.Age).To(Equal(uint(29)))
 	})
@@ -220,7 +239,7 @@ var _ = Describe("ParseArgs", func() {
 		var args struct {
 			Age uint64 `arg:"age"`
 		}
-		err := argument.ParseArgs(&args, []string{"-age=29"})
+		err := argument.ParseArgs(ctx, &args, []string{"-age=29"})
 		Expect(err).NotTo(HaveOccurred())
 		Expect(args.Age).To(Equal(uint64(29)))
 	})
@@ -228,7 +247,7 @@ var _ = Describe("ParseArgs", func() {
 		var args struct {
 			Age uint64 `arg:"age" default:"29"`
 		}
-		err := argument.ParseArgs(&args, []string{})
+		err := argument.ParseArgs(ctx, &args, []string{})
 		Expect(err).NotTo(HaveOccurred())
 		Expect(args.Age).To(Equal(uint64(29)))
 	})
@@ -236,14 +255,14 @@ var _ = Describe("ParseArgs", func() {
 		var args struct {
 			Age interface{} `arg:"age" default:"29"`
 		}
-		err := argument.ParseArgs(&args, []string{})
+		err := argument.ParseArgs(ctx, &args, []string{})
 		Expect(err).To(HaveOccurred())
 	})
 	It("parse int32", func() {
 		var args struct {
 			Age int32 `arg:"age"`
 		}
-		err := argument.ParseArgs(&args, []string{"-age=29"})
+		err := argument.ParseArgs(ctx, &args, []string{"-age=29"})
 		Expect(err).NotTo(HaveOccurred())
 		Expect(args.Age).To(Equal(int32(29)))
 	})
@@ -251,7 +270,7 @@ var _ = Describe("ParseArgs", func() {
 		var args struct {
 			Age int32 `arg:"age" default:"29"`
 		}
-		err := argument.ParseArgs(&args, []string{})
+		err := argument.ParseArgs(ctx, &args, []string{})
 		Expect(err).NotTo(HaveOccurred())
 		Expect(args.Age).To(Equal(int32(29)))
 	})