+++ env_test.go
@@ -5,6 +5,7 @@
 package argument_test
 
 import (
+	"context"
 	"time"
 
 	. "github.com/onsi/ginkgo"
@@ -14,11 +15,15 @@ import (
 )
 
 var _ = Describe("ParseEnv", func() {
+	var ctx context.Context
+	BeforeEach(func() {
+		ctx = context.Background()
+	})
 	It("parse string", func() {
 		var args struct {
 			Username string `env:"user"`
 		}
-		err := argument.ParseEnv(&args, []string{"user=Ben"})
+		err := argument.ParseEnv(ctx, &args, []string{"user=Ben"})
 		Expect(err).NotTo(HaveOccurred())
 		Expect(args.Username).To(Equal("Ben"))
 	})
@@ -26,7 +31,7 @@ var _ = Describe("ParseEnv", func() {
 		var args struct {
 			Age int `env:"age"`
 		}
-		err := argument.ParseEnv(&args, []string{"age=29"})
+		err := argument.ParseEnv(ctx, &args, []string{"age=29"})
 		Expect(err).NotTo(HaveOccurred())
 		Expect(args.Age).To(Equal(29))
 	})
@@ -34,21 +39,21 @@ var _ = Describe("ParseEnv", func() {
 		var args struct {
 			Age int `env:"age"`
 		}
-		err := argument.ParseEnv(&args, []string{"age=abc"})
+		err := argument.ParseEnv(ctx, &args, []string{"age=abc"})
 		Expect(err).To(HaveOccurred())
 	})
 	It("return error if parse int64 fails", func() {
 		var args struct {
 			Age int64 `env:"age"`
 		}
-		err := argument.ParseEnv(&args, []string{"age=abc"})
+		err := argument.ParseEnv(ctx, &args, []string{"age=abc"})
 		Expect(err).To(HaveOccurred())
 	})
 	It("parse int64", func() {
 		var args struct {
 			Age int64 `env:"age"`
 		}
-		err := argument.ParseEnv(&args, []string{"age=29"})
+		err := argument.ParseEnv(ctx, &args, []string{"age=29"})
 		Expect(err).NotTo(HaveOccurred())
 		Expect(args.Age).To(Equal(int64(29)))
 	})
@@ -56,7 +61,7 @@ var _ = Describe("ParseEnv", func() {
 		var args struct {
 			Confirm bool `env:"confirm"`
 		}
-		err := argument.ParseEnv(&args, []string{"confirm=true"})
+		err := argument.ParseEnv(ctx, &args, []string{"confirm=true"})
 		Expect(err).NotTo(HaveOccurred())
 		Expect(args.Confirm).To(BeTrue())
 	})
@@ -64,7 +69,7 @@ var _ = Describe("ParseEnv", func() {
 		var args struct {
 			Confirm bool `env:"confirm"`
 		}
-		err := argument.ParseEnv(&args, []string{"confirm=false"})
+		err := argument.ParseEnv(ctx, &args, []string{"confirm=false"})
 		Expect(err).NotTo(HaveOccurred())
 		Expect(args.Confirm).To(BeFalse())
 	})
@@ -72,14 +77,14 @@ var _ = Describe("ParseEnv", func() {
 		var args struct {
 			Confirm bool `env:"confirm"`
 		}
-		err := argument.ParseEnv(&args, []string{"confirm=hello"})
+		err := argument.ParseEnv(ctx, &args, []string{"confirm=hello"})
 		Expect(err).To(HaveOccurred())
 	})
 	It("parse duration", func() {
 		var args struct {
 			Wait time.Duration `env:"wait"`
 		}
-		err := argument.ParseEnv(&args, []string{"wait=1m"})
+		err := argument.ParseEnv(ctx, &args, []string{"wait=1m"})
 		Expect(err).NotTo(HaveOccurred())
 		Expect(args.Wait).To(Equal(time.Minute))
 	})
@@ -87,14 +92,14 @@ var _ = Describe("ParseEnv", func() {
 		var args struct {
 			Wait time.Duration `env:"wait"`
 		}
-		err := argument.ParseEnv(&args, []string{"wait=hello"})
+		err := argument.ParseEnv(ctx, &args, []string{"wait=hello"})
 		Expect(err).To(HaveOccurred())
 	})
 	It("parse float64", func() {
 		var args struct {
 			Age float64 `env:"age"`
 		}
-		err := argument.ParseEnv(&args, []string{"age=29"})
+		err := argument.ParseEnv(ctx, &args, []string{"age=29"})
 		Expect(err).NotTo(HaveOccurred())
 		Expect(args.Age).To(Equal(float64(29)))
 	})
@@ -102,14 +107,14 @@ var _ = Describe("ParseEnv", func() {
 		var args struct {
 			Age float64 `env:"age"`
 		}
-		err := argument.ParseEnv(&args, []string{"age=abc"})
+		err := argument.ParseEnv(ctx, &args, []string{"age=abc"})
 		Expect(err).To(HaveOccurred())
 	})
 	It("parse uint", func() {
 		var args struct {
 			Age uint `env:"age"`
 		}
-		err := argument.ParseEnv(&args, []string{"age=29"})
+		err := argument.ParseEnv(ctx, &args, []string{"age=29"})
 		Expect(err).NotTo(HaveOccurred())
 		Expect(args.Age).To(Equal(uint(29)))
 	})
@@ -117,14 +122,14 @@ var _ = Describe("ParseEnv", func() {
 		var args struct {
 			Age uint `env:"age"`
 		}
-		err := argument.ParseEnv(&args, []string{"age=abc"})
+		err := argument.ParseEnv(ctx, &args, []string{"age=abc"})
 		Expect(err).To(HaveOccurred())
 	})
 	It("parse uint64", func() {
 		var args struct {
 			Age uint64 `env:"age"`
 		}
-		err := argument.ParseEnv(&args, []string{"age=29"})
+		err := argument.ParseEnv(ctx, &args, []string{"age=29"})
 		Expect(err).NotTo(HaveOccurred())
 		Expect(args.Age).To(Equal(uint64(29)))
 	})
@@ -132,14 +137,14 @@ var _ = Describe("ParseEnv", func() {
 		var args struct {
 			Age uint64 `env:"age"`
 		}
-		err := argument.ParseEnv(&args, []string{"age=abc"})
+		err := argument.ParseEnv(ctx, &args, []string{"age=abc"})
 		Expect(err).To(HaveOccurred())
 	})
 	It("skip fields without tag", func() {
 		var args struct {
 			Age int
 		}
-		err := argument.ParseEnv(&args, []string{})
+		err := argument.ParseEnv(ctx, &args, []string{})
 		Expect(err).NotTo(HaveOccurred())
 		Expect(args.Age).To(Equal(0))
 	})
@@ -147,14 +152,14 @@ var _ = Describe("ParseEnv", func() {
 		var args struct {
 			Age interface{} `env:"age" default:"29"`
 		}
-		err := argument.ParseEnv(&args, []string{"age=abc"})
+		err := argument.ParseEnv(ctx, &args, []string{"age=abc"})
 		Expect(err).NotTo(BeNil())
 	})
 	It("parse int32", func() {
 		var args struct {
 			Age int32 `env:"age"`
 		}
-		err := argument.ParseEnv(&args, []string{"age=29"})
+		err := argument.ParseEnv(ctx, &args, []string{"age=29"})
 		Expect(err).NotTo(HaveOccurred())
 		Expect(args.Age).To(Equal(int32(29)))
 	})
@@ -162,7 +167,7 @@ var _ = Describe("ParseEnv", func() {
 		var args struct {
 			Age int32 `env:"age"`
 		}
-		err := argument.ParseEnv(&args, []string{"age=abc"})
+		err := argument.ParseEnv(ctx, &args, []string{"age=abc"})
 		Expect(err).To(HaveOccurred())
 	})
 })