+++ CHANGELOG.md
@@ -8,6 +8,11 @@ Please choose versions by [Semantic Versioning](http://semver.org/).
 * MINOR version when you add functionality in a backwards-compatible manner, and
 * PATCH version when you make backwards-compatible bug fixes.
 
+## v2.0.0
+
+- add context to all methods
+- allow parse duration with unit d=day and w=week
+
 ## v1.3.2
 
 - fix int32